pyproject.toml
[project]
name = "video-editing-agent"
version = "0.1.0"
description = "Multi-agent AI video editor powered by CrewAI and Ollama (fully local, no API keys)"
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.108.0",
    "openai>=1.0",
    "requests>=2.31.0",
    "ffmpeg-python>=0.2.0",
    "opencv-python>=4.8.0",
    "Pillow>=10.0",
    "streamlit>=1.40",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
    "ruff>=0.4",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
